improved edge detection design
